git-svn-id: http://svn.seqan.de/seqan/trunk@4781 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/misc/misc_cmdparser.h
@@ -1201,7 +1201,7 @@ inline bool
 _convertOptionValue(CommandLineOption const & opt, float & dst, CharString const & src)
 {
     if (!isDoubleOption(opt)) return false;
-	dst = atof(toCString(src));
+	dst = (float)atof(toCString(src));
 	return length(src) > 0;
 }
